Split CheckoutForm into smaller components for each step
