+++ assets/js/blocks/cart-checkout/checkout/form/order-notes-step.js
@@ -1,0 +1,56 @@
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+import { FormStep } from '@woocommerce/base-components/cart-checkout';
+import {
+	useCheckoutContext,
+	useShippingDataContext,
+} from '@woocommerce/base-context';
+import PropTypes from 'prop-types';
+
+/**
+ * Internal dependencies
+ */
+import CheckoutOrderNotes from './order-notes';
+
+const OrderNotesStep = ( { showOrderNotes } ) => {
+	const { needsShipping } = useShippingDataContext();
+	const {
+		isProcessing: checkoutIsProcessing,
+		orderNotes,
+		dispatchActions,
+	} = useCheckoutContext();
+	const { setOrderNotes } = dispatchActions;
+
+	if ( ! showOrderNotes ) {
+		return null;
+	}
+
+	return (
+		<FormStep id="order-notes" showStepNumber={ false }>
+			<CheckoutOrderNotes
+				disabled={ checkoutIsProcessing }
+				onChange={ setOrderNotes }
+				placeholder={
+					needsShipping
+						? __(
+								'Notes about your order, e.g. special notes for delivery.',
+								'woo-gutenberg-products-block'
+						  )
+						: __(
+								'Notes about your order.',
+								'woo-gutenberg-products-block'
+						  )
+				}
+				value={ orderNotes }
+			/>
+		</FormStep>
+	);
+};
+
+OrderNotesStep.propTypes = {
+	showOrderNotes: PropTypes.bool.isRequired,
+};
+
+export default OrderNotesStep;
